Backend-python/requirements.txt
flask
flask-cors
pandas
scikit-learn
xgboost
joblib
python-dotenv
httpx
nltk

# Local HuggingFace AI (already installed in venv)
transformers
langchain
langchain_huggingface
torch
sentencepiece

# Optional — chart image generation (not used in current API)
# seaborn
# matplotlib
